+++ vj4/ui/templates/ranking.html
@@ -1,10 +1,10 @@
 {% extends "layout/basic.html" %}
 {% block content %}
 <div class="row">
-  <div class="medium-9 columns">
+  <div class="medium-12 columns">
     <div class="section">
       <div class="section__body no-padding">
-        {% include "partials/rank.html" %}
+        {% include "partials/ranking.html" %}
       </div>
     </div>
   </div>
